UIControl+addTouchAnimation extracted
