+++ KarhooUISDK/CommonUI/Views/MainActionButton.swift
@@ -38,10 +38,6 @@ class MainActionButton: UIButton {
         setTitleColor(KarhooUI.colors.white, for: .normal)
         
         addTouchAnimation()
-//        addTarget(self, action: #selector(pressed), for: .touchDown)
-//        addTarget(self, action: #selector(touchCancelled), for: .touchCancel)
-//        addTarget(self, action: #selector(released), for: .touchUpOutside)
-//        addTarget(self, action: #selector(released), for: .touchUpInside)
     }
 
     private func setupLayout() {
@@ -52,27 +48,6 @@ class MainActionButton: UIButton {
         }.isActive = true
     }
 
-//    @objc
-//    private func pressed() {
-//        UIView.animate(
-//            withDuration: UIConstants.Duration.xShort,
-//            delay: 0,
-//            options: .curveEaseOut,
-//            animations: { [weak self] in
-//                self?.transform = UIConstants.Dimension.View.mainActionButtonPressedAffineTransform
-//            }
-//        )
-//    }
-
-//    @objc
-//    private func released() {
-//        animateToDefaultState()
-//    }
-//
-//    @objc private func touchCancelled() {
-//        animateToDefaultState()
-//    }
-
     private func animateToDefaultState() {
         UIView.animate(
             withDuration: UIConstants.Duration.xShort,
@@ -80,48 +55,6 @@ class MainActionButton: UIButton {
             options: .curveEaseOut,
             animations: { [weak self] in
                 self?.transform = .identity
-            }
-        )
-    }
-}
-
-extension UIControl {
-    func addTouchAnimation() {
-        addTarget(self, action: #selector(pressed), for: .touchDown)
-        addTarget(self, action: #selector(touchCancelled), for: .touchCancel)
-        addTarget(self, action: #selector(released), for: .touchUpOutside)
-        addTarget(self, action: #selector(released), for: .touchUpInside)
-    }
-
-    @objc
-    private func released() {
-        animateToDefaultState()
-    }
-
-    @objc
-    private func touchCancelled() {
-        animateToDefaultState()
-    }
-
-    private func animateToDefaultState() {
-        UIView.animate(
-            withDuration: UIConstants.Duration.xShort,
-            delay: 0,
-            options: .curveEaseOut,
-            animations: { [weak self] in
-                self?.transform = .identity
-            }
-        )
-    }
-
-    @objc
-    private func pressed() {
-        UIView.animate(
-            withDuration: UIConstants.Duration.xShort,
-            delay: 0,
-            options: .curveEaseOut,
-            animations: { [weak self] in
-                self?.transform = UIConstants.Dimension.View.mainActionButtonPressedAffineTransform
             }
         )
     }
